Switching time calculations to double to eliminate NaNs

+++ examples/fdtd/src/defines.hpp
@@ -46,7 +46,7 @@ constexpr float dx = 10 * nm;
 constexpr float sqrt_2 = 1.4142135623730951;
 
 // distance between generations in s
-constexpr float dt = (dx / (c0 * sqrt_2)) * 0.99;
+constexpr double dt = (double(dx) / double(c0 * sqrt_2)) * 0.99;
 
 // Number of values in a vector cell.
 constexpr uindex_t vector_len = 8;
@@ -65,8 +65,8 @@ constexpr uindex_t pipeline_length = 16;
 
 // Default radius of the cavity in dx.
 constexpr float default_radius = 80.0;
-constexpr float default_tau = 100e-15;
-constexpr float default_frequency = 121.5e12;
+constexpr double default_tau = 100e-15;
+constexpr double default_frequency = 121.5e12;
 
 struct FDTDCell
 {
@@ -120,31 +120,31 @@ struct Parameters
     float disk_radius;
 
     // Timescale (?) for the source wave in s.
-    float tau() const
+    double tau() const
     {
         return default_tau * (disk_radius / default_radius);
     }
 
     // The frequency of the source wave in Hz.
-    float frequency() const
+    double frequency() const
     {
         return default_frequency * (default_radius / disk_radius);
     }
 
     // Omega (?) in Hz.
-    float omega() const
+    double omega() const
     {
         return 2.0 * M_PI * frequency();
     }
 
     // The point in time when the simulation starts in s.
-    float t0() const
+    double t0() const
     {
         return 3.0 * tau();
     }
 
     // The point in time to cut off the source wave in s.
-    float t_cutoff() const
+    double t_cutoff() const
     {
         return 7.0 * tau();
     }
