+++ examples/fdtd/src/simulation.hpp
@@ -32,14 +32,14 @@ constexpr static Material vacuum{
 class FDTDKernel
 {
     float disk_radius;
-    float tau;
-    float omega;
-    float t0;
-    float t_cutoff;
-    float t_passed;
+    double tau;
+    double omega;
+    double t0;
+    double t_cutoff;
+    double t_passed;
 
 public:
-    FDTDKernel(Parameters const &parameters, float t_passed) : disk_radius(parameters.disk_radius), tau(parameters.tau()), omega(parameters.omega()), t0(parameters.t0()), t_cutoff(parameters.t_cutoff()), t_passed(t_passed) {}
+    FDTDKernel(Parameters const &parameters, double t_passed) : disk_radius(parameters.disk_radius), tau(parameters.tau()), omega(parameters.omega()), t0(parameters.t0()), t_cutoff(parameters.t_cutoff()), t_passed(t_passed) {}
 
     static FDTDCell halo()
     {
@@ -102,11 +102,11 @@ public:
             }
             top_neighbours[0] = stencil[ID(0, -1)].hz[vector_len - 1];
 
-            cell.ex *= ca;                                           // 8 FOs
-            cell.ex += cb * (stencil[ID(0, 0)].hz - top_neighbours); // 8*2 = 16 FOs
+            cell.ex *= ca;
+            cell.ex += cb * (stencil[ID(0, 0)].hz - top_neighbours);
 
-            cell.ey *= ca;                                            // 8FOs
-            cell.ey += cb * (left_neighbours - stencil[ID(0, 0)].hz); // 8*2 = 16 FOs
+            cell.ey *= ca;
+            cell.ey += cb * (left_neighbours - stencil[ID(0, 0)].hz);
         }
         else
         {
@@ -120,14 +120,14 @@ public:
             }
             bottom_neighbours[vector_len - 1] = stencil[ID(0, 1)].ex[0];
 
-            cell.hz *= da;                                                                                        // 8 FOs
-            cell.hz += db * (bottom_neighbours - stencil[ID(0, 0)].ex + stencil[ID(0, 0)].ey - right_neighbours); // 8*4 = 32 FOs
+            cell.hz *= da;
+            cell.hz += db * (bottom_neighbours - stencil[ID(0, 0)].ex + stencil[ID(0, 0)].ey - right_neighbours);
 
-            float current_time = t_passed + (stencil.generation >> 1) * dt; // 1 FO
+            double current_time = t_passed + (stencil.generation >> 1) * dt;
             if (current_time < t_cutoff)
             {
-                float wave_progress = (current_time - t0) / tau;                                                    // 3 FOs
-                cell.hz += cl::sycl::cos(omega * current_time) * cl::sycl::exp(-1 * wave_progress * wave_progress); // 8 + 6 FOs
+                double wave_progress = (current_time - t0) / tau;
+                cell.hz += cl::sycl::cos(omega * current_time) * cl::sycl::exp(-1 * wave_progress * wave_progress);
             }
 
             cell.hz_sum += cell.hz * cell.hz; // 2*8 = 16 FOs
